refactor: 승자 계산과 수익 계산을 Participants에 위임
