+++ src/main/java/domain/BlackjackGame.java
@@ -5,7 +5,6 @@ import domain.participant.Dealer;
 import domain.participant.Participant;
 import domain.participant.Participants;
 import domain.participant.Player;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -47,42 +46,8 @@ public class BlackjackGame {
     }
 
     public Map<Player, Integer> getBettingResult() {
-        Map<Player, Integer> betResult = new LinkedHashMap<>();
-        for (Player player : participants.getPlayers()) {
-            PlayerGameResult playerGameResult = getPlayerGameResult(player);
-            int reward = playerGameResult.calculateBenefit(player.getBetAmount());
-            betResult.put(player, reward);
-        }
-
-        return new LinkedHashMap<>(betResult);
+        return participants.getBettingResult();
     }
-
-    private PlayerGameResult getPlayerGameResult(Player player) {
-        if (isPlayerBlackjack(player)) {
-            return PlayerGameResult.BLACKJACK;
-        }
-        if (isPlayerWin(player)) {
-            return PlayerGameResult.WIN;
-        }
-        if (isDraw(player)) {
-            return PlayerGameResult.DRAW;
-        }
-        return PlayerGameResult.LOSE;
-    }
-
-    private boolean isPlayerBlackjack(Player player) {
-        return player.isBlackjack() && !participants.isDealerBlackjack();
-    }
-
-    private boolean isPlayerWin(Player player) {
-        return (player.calculateScore() > participants.getDealerScore() && !player.isBust())
-                || participants.isDealerBust();
-    }
-
-    private boolean isDraw(Player player) {
-        return player.calculateScore() == participants.getDealerScore();
-    }
-
 
     public Dealer getDealer() {
         return participants.getDealer();
